refactor: 매물 리스트 컴포넌트 세분화
