+++ src/App.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import LoginForm from './components/Login/LoginForm';
 import HeaderForm from './components/Header/HeaderForm';
 import SideMenuForm from './components/SideMenu/SideMenuForm';
+import ForSaleListForm from './components/Market/ForSaleList/ForSaleListForm';
 import VehicleRegistrationInquriyForm from './components/Market/VehicleRegistrationInquiry/VehicleRegistrationInquriyForm';
 import VehicleRegistrationInquiryDetailForm from './components/Market/VehicleRegistrationInquiryDetail/VehicleRegistrationInquiryDetailForm';
 
@@ -15,9 +16,9 @@ function App() {
           <HeaderForm />
           <div style={{ display: "flex", height:"100%" }}>
             <SideMenuForm />
-            {/* <ForSaleListForm /> */}
+            <ForSaleListForm />
             {/* <ForSaleListDetailForm /> */}
-            <VehicleRegistrationInquriyForm />
+            {/* <VehicleRegistrationInquriyForm /> */}
             {/* <VehicleRegistrationInquiryDetailForm /> */}
           </div>
         </>
