+++ src/components/Market/ForSaleListDetail/ForSaleListDetailForm.styled.ts
@@ -15,6 +15,7 @@ export const ContentWrapperAttrs = styled.div`
 `;
 
 export const TableAttrs = styled.table`
+    width: 100%;
     font-size: ${({theme}) => theme.fontSize.header_14};
     font-weight: ${({theme}) => theme.fontWeight.normal};
     line-height: 150%;
