+++ src/components/Market/ReadMe.txt
@@ -1,3 +1,28 @@
 Author      : 지용
-Date        : 2022-12-17
+Date        : 2022-12-22
 Comments    : Market 디렉토리 설명
+
+├─Detail                                    : 각 상세 페이지에 쓰일 공용 컴포넌트 디렉토리
+│  ├─DownloadFile                           : 첨부파일 다운로드 항목 컴포넌트 모음
+│  ├─MainInfoTable                          : 차량정보, 차종 등 정보 테이블
+│  ├─PhotoRegistrationTable                 : 사진 등록 보이기 테이블
+│  ├─SellerInfoTable                        : 판매자 정보 테이블
+│  │  ├─DealerInfoTable                     : 판매자 정보 -> 딜러 정보 테이블
+│  │  └─MemberInfoTable                     : 판매자 정보 -> 회원 정보 테이블
+│  ├─SubTitleTable                          : 부제목 테이블
+│  └─VehicleDetailTable                     : 차량 설명글
+├─ForSaleList
+│  ├─ForSaleMainTable
+│  └─PageItems
+│      ├─Table
+│      │  ├─Body
+│      │  └─Header
+│      └─TopBanner
+├─ForSaleListDetail                         : 매물리스트 상세보기
+├─VehicleRegistrationInquiry                : 판매차량 등록 문의 리스트
+│  └─PageItems                              : 판매차량 등록 문의 리스트 페이지 아이템
+│    ├─Table                                : 테이블 아이템 모음
+│    │  ├─Body                              : 테이블 바디
+│    │  └─Header                            : 테이블 헤더
+│    └─TopBanner                            : Top banner, dropdown, search bar 모음
+└─VehicleRegistrationInquiryDetail          : 판매차량 등록 문의 리스트 상세 보기
